Huiswerk Week 7

+++ INOOP_Week7_OpenbaarVervoer/INOOP_Week7_OpenbaarVervoer/Metro.h
@@ -8,7 +8,8 @@ public:
 	Metro(std::string vervoersmiddel);
 	void prijsPerKm();
 	void maxPassagiers();
-	int prijs;
+
+	float prijs;
 	int passagiers;
 
 };
